0-linear.c

+++ 0x1E-search_algorithms/0-linear.c
@@ -9,13 +9,13 @@
  **/
 int linear_search(int *array, size_t size, int value)
 {
-	int i = 0;
+	size_t i = 0;
 
 	if (array == NULL)
 		return (-1);
 	while (i < size)
 	{
-		printf("Value checked array[%i] = [%i]\n", i, array[i]);
+		printf("Value checked array[%li] = [%i]\n", i, array[i]);
 		if (array[i] == value)
 			return (i);
 		i++;
